+++ src/hdf/object/nc2/NC2Group.java
@@ -29,7 +29,7 @@ import ucar.nc2.NetcdfFile;
  * references to other groups or data objects just as the UNIX directory may
  * contain subdirectories or files.
  * <p>
- * 
+ *
  * @version 1.1 9/4/2007
  * @author Peter X. Cao
  */
@@ -48,7 +48,7 @@ public class NC2Group extends Group {
     /**
      * Constructs an HDF5 group with specific name, path, and parent.
      * <p>
-     * 
+     *
      * @param fileFormat
      *            the file which containing the group.
      * @param name
@@ -68,7 +68,7 @@ public class NC2Group extends Group {
 
     /*
      * (non-Javadoc)
-     * 
+     *
      * @see hdf.object.DataFormat#hasAttribute()
      */
     public boolean hasAttribute() {
@@ -108,9 +108,9 @@ public class NC2Group extends Group {
     /**
      * Creates a new attribute and attached to this dataset if attribute does
      * not exist. Otherwise, just update the value of the attribute.
-     * 
+     *
      * <p>
-     * 
+     *
      * @param info
      *            the atribute to attach
      */
@@ -123,7 +123,7 @@ public class NC2Group extends Group {
     /**
      * Deletes an attribute from this dataset.
      * <p>
-     * 
+     *
      * @param info
      *            the attribute to delete.
      */
@@ -141,7 +141,7 @@ public class NC2Group extends Group {
 
     // Implementing DataFormat
     @Override
-    public int open() {
+    public long open() {
         // not supported
         throw new UnsupportedOperationException(
                 "Unsupported operation for NetCDF.");
@@ -149,7 +149,7 @@ public class NC2Group extends Group {
 
     /** close group access */
     @Override
-    public void close(int gid) {
+    public void close(long gid) {
         // not supported
         throw new UnsupportedOperationException(
                 "Unsupported operation for NetCDF.");
@@ -157,7 +157,7 @@ public class NC2Group extends Group {
 
     /**
      * Creates a new group.
-     * 
+     *
      * @param file
      *            the file which the group is added to.
      * @param name
